# docs/sydney-movies-app-openapi.yaml
openapi: 3.0.3
info:
  title: Sydney Movies App API
  version: 1.0.0
  description: |
    REST API for the Sydney Movies App (Your Cine Stacks). The API mediates all
    access to the underlying PostgreSQL/Supabase database. The client must
    never connect directly to the database.

    Key responsibilities:
    - User profile & preferences
    - Movies catalog (read-heavy, optional admin CRUD)
    - Watchlists & items
    - Genre preferences
    - Weekly recommendations
    - Smart filtering & search via query parameters on /movies
servers:
  - url: https://api.sydney-movies.test/v1
    description: Local/dev server (example)
...
paths:
...
  /movies:
    get:
      tags: [Movies]
      summary: List & search movies
      description: |
        Returns a paginated list of movies. Supports keyword search and
        smart filtering by genre, rating, language, runtime, and sorting
        (used to power the "smart filtering" and search stories). :contentReference[oaicite:2]{index=2}
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          schema:
            type: string
          description: Keyword search on title (case-insensitive).
        - in: query
          name: genreId
          schema:
            type: integer
            format: int64
          description: Filter by genre ID.
        - in: query
          name: ageRating
          schema:
            type: string
          description: Filter by MPAA-style age rating (e.g. G, PG, PG-13, R).
        - in: query
          name: language
          schema:
            type: string
          description: Filter by original language (e.g. "en", "ja").
        - in: query
          name: runtimeMin
          schema:
            type: integer
          description: Minimum runtime in minutes.
        - in: query
          name: runtimeMax
          schema:
            type: integer
          description: Maximum runtime in minutes.
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [title, release_year, average_viewer_rating, created_at]
          description: Sort by field.
        - in: query
          name: sortDir
          schema:
            type: string
            enum: [asc, desc]
          description: Sort direction (asc or desc).
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
          description: Page size.
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          description: Offset for pagination.
      responses:
        '200':
          description: List of movies.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Movie'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags: [Movies]
      summary: Create a new movie (admin/internal)
      description: |
        Create a new movie record in the catalog. In production, this would
        typically be done by an admin service or via TMDB/IMDb sync, not by
        end-users.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieCreateInput'
      responses:
        '201':
          description: Movie created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /genres:
    get:
      tags: [Genres]
      summary: List all genres
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of genres.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Genre'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags: [Genres]
      summary: Create a new genre (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenreCreateInput'
      responses:
        '201':
          description: Genre created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Genre'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /watchlists:
    get:
      tags: [Watchlists]
      summary: List current user's watchlists
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of watchlists owned by the current user.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Watchlist'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags: [Watchlists]
      summary: Create a new watchlist
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WatchlistCreateInput'
      responses:
        '201':
          description: Watchlist created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Watchlist'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Supabase JWT access token. The backend validates the token and
        uses the user ID (auth.uid) to scope profile and user-owned data.
...
  responses:
    BadRequest:
      description: Bad request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UnauthorizedError:
      description: Authentication failed or missing.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Movie:
      type: object
      properties:
        movie_id:
          type: integer
          format: int64
          readOnly: true
        title:
          type: string
        release_year:
          type: integer
          nullable: true
        age_rating:
          type: string
          nullable: true
        runtime_minutes:
          type: integer
          nullable: true
        original_language:
          type: string
          nullable: true
        average_viewer_rating:
          type: number
          format: double
          description: Aggregated average (0–100) based on user ratings.
        poster_url:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required: [title]
    MovieCreateInput:
      type: object
      properties:
        title:
          type: string
        release_year:
          type: integer
        age_rating:
          type: string
        runtime_minutes:
          type: integer
        original_language:
          type: string
        poster_url:
          type: string
      required: [title]
...
    Genre:
      type: object
      properties:
        genre_id:
          type: integer
          format: int64
          readOnly: true
        name:
          type: string
        created_at:
          type: string
          format: date-time
          readOnly: true
      required: [name]
    GenreCreateInput:
      type: object
      properties:
        name:
          type: string
      required: [name]
...
    Watchlist:
      type: object
      properties:
        watchlist_id:
          type: integer
          format: int64
          readOnly: true
        user_id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required: [name]
    WatchlistCreateInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required: [name]
...
    Error:
      type: object
      properties:
        code:
          type: string
          example: bad_request
        message:
          type: string
        details:
          type: object
          additionalProperties: true